+++ src/main/java/utils/PropertyReader.java
@@ -1,4 +1,4 @@
-package com.qabot.utils;
+package utils;
 
 import java.io.IOException;
 import java.io.InputStream;
