+++ src/test/java/com/qabot/MasterTest.java
@@ -1,16 +1,46 @@
 package com.qabot;
 
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.annotations.*;
+
+import java.io.FileInputStream;
 import java.io.IOException;
+import java.time.Duration;
+import java.util.Properties;
 
 
-public class MasterTest extends MasterPage{
+public class MasterTest {
 
-    @Test
-    public void basePageNavigation() throws IOException {
-        driver = initializeDriver();
-        driver.get("http://the-internet.herokuapp.com/");
+    public WebDriver driver;
 
+    public WebDriver initializeDriver() throws IOException {
+
+        // Chrome driver
+        Properties prop = new Properties();
+        FileInputStream file = new FileInputStream(System.getProperty("user.dir") + "/properties/userdata.properties");
+        prop.load(file);
+        String browserName = prop.getProperty("browser");
+        System.out.println(browserName);
+
+        if (browserName.equals("chrome")) {
+            System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/drivers/chromedriver.exe");
+            driver = new ChromeDriver();
+
+        } else if (browserName.equals("firefox")) {
+            // Firefox code
+            System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/drivers/chromedriver.exe");
+            driver = new ChromeDriver();
+
+        } else if (browserName.equals("IE")) {
+            // IE code
+            System.setProperty("webdriver.chrome.driver", System.getProperty("user.dir") + "/drivers/chromedriver.exe");
+            driver = new ChromeDriver();
+        }
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+        return driver;
     }
 
 }
+
+
